requirements.txt
PyYAML~=6.0
setuptools~=65.5.1